Use generic status error for unexpected http errors
Removes the text about expected error code in favor of a more generic
http status error. Also pulls shared error-related code into a separate
file.
